conversion script

+++ catalogs/obs/catalog/ERA5/scripts/reprocess-rename.sh
@@ -4,11 +4,15 @@ year1=1940
 year2=2025
 
 mkdir -p $DIRNEW
-for file in $DIR/ERA5_*_mon_full_sfc_${year1}-${year2}.nc ; do
+for file in $DIR/ERA5_*_mon_full_*_${year1}-${year2}.nc ; do
     filename=$(basename $file)
     echo $filename
 
     rm -f $DIRNEW/test.grb
+    if [ -f $DIRNEW/$filename ]; then
+        echo "File $DIRNEW/$filename already exists. Skipping."
+        continue
+    fi
     cdo -f grb copy $DIR/$filename $DIRNEW/test.grb
     cdo -f nc4 -z zip --eccodes copy $DIRNEW/test.grb $DIRNEW/$filename
     rm $DIRNEW/test.grb
